Proper daemon support
