+++ init/gns3.conf.upstart
@@ -5,8 +5,7 @@ start on filesystem or runlevel [2345]
 stop on shutdown
 
 script
-    echo $$ > /var/run/gns3.pid
-    exec start-stop-daemon --start -c gns3 --exec /usr/local/bin/gns3server --log /var/log/gns3.log
+    exec start-stop-daemon --start -c gns3 --exec /usr/local/bin/gns3server --log /var/log/gns3.log --pid /var/run/gns3.pid --daemon
 end script
 
 pre-start script
@@ -14,6 +13,5 @@ pre-start script
 end script
 
 pre-stop script
-    rm /var/run/gns3.pid
     echo "[`date`] GNS3 Stopping" >> /var/log/gns3.log
 end script
